Made stricter checks for types
