+++ google_docstring_parser/__init__.py
@@ -4,5 +4,6 @@ A lightweight, efficient parser for Google-style Python docstrings that converts
 """
 
 from google_docstring_parser.google_docstring_parser import ReferenceFormatError, parse_google_docstring
+from google_docstring_parser.type_validation import InvalidTypeAnnotationError
 
-__all__ = ["ReferenceFormatError", "parse_google_docstring"]
+__all__ = ["InvalidTypeAnnotationError", "ReferenceFormatError", "parse_google_docstring"]
